- UserRepositoryImpl.kt - reauthenticate 함수를 callbackFlow 로 변경 - cancelSignin 의 내부를 동기로 작동하도록 변경

+++ data/src/main/java/com/gyleedev/data/repository/UserRepositoryImpl.kt
@@ -10,6 +10,7 @@ import androidx.paging.PagingData
 import androidx.paging.map
 import com.google.firebase.auth.EmailAuthProvider
 import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.FirebaseUser
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.FirebaseDatabase
@@ -885,15 +886,26 @@ class UserRepositoryImpl @Inject constructor(
         return cancelResult
     }
 
-    private fun reauthenticate() {
+    private fun reauthenticate(): Flow<Boolean> = callbackFlow {
         val user = auth.currentUser!!
         val credential = EmailAuthProvider
-            .getCredential(myDataPreference.getMyData().email, authenticationPreference.getPassword())
-        user.reauthenticate(credential)
+            .getCredential(
+                myDataPreference.getMyData().email,
+                authenticationPreference.getPassword()
+            )
+        user.reauthenticate(credential).addOnCompleteListener { task ->
+            trySend(task.isSuccessful)
+        }
+        awaitClose()
     }
 
     private fun cancelSignin(): Flow<Boolean> = callbackFlow {
-        reauthenticate()
+        val reauthenticateSuccess = reauthenticate().first()
+        if (!reauthenticateSuccess) {
+            trySend(false)
+            awaitClose()
+            return@callbackFlow
+        }
         auth.currentUser!!.delete().addOnCompleteListener { task ->
             if (task.isSuccessful) {
                 auth.signOut()
@@ -924,6 +936,7 @@ class UserRepositoryImpl @Inject constructor(
     override suspend fun setVerifiedState(verifiedState: VerifiedState) {
         authenticationPreference.setState(verifiedState)
     }
+
     override suspend fun verifyEmailRequest(): Flow<Boolean> = callbackFlow {
         requireNotNull(auth.currentUser).sendEmailVerification()
             .addOnCompleteListener { task ->
